# cmd/api/docs/swagger.yaml
basePath: /
definitions:
...
  dto.CreateOtherRequest:
    properties:
      active:
        example: true
        type: boolean
      customer_id:
        type: integer
      key:
        example: company_size
        type: string
      value:
        example: 50-100 employees
        type: string
    required:
    - customer_id
    - key
    type: object
...
  dto.ErrorResponse:
    properties:
      error:
        example: Invalid request
        type: string
    type: object
...
  entity.Activity:
    properties:
      activity_checkins:
        items:
          $ref: '#/definitions/entity.ActivityCheckin'
        type: array
      agenda:
        type: string
      attendees:
        items:
          $ref: '#/definitions/entity.User'
        type: array
      created_at:
        type: string
      created_by:
        type: integer
      creator:
        $ref: '#/definitions/entity.User'
      customer_id:
        type: integer
      end_time:
        type: string
      id:
        type: integer
      location_name:
        type: string
      start_time:
        type: string
      status:
        type: string
      title:
        type: string
      type:
        type: string
      updated_at:
        type: string
    type: object
  entity.ActivityCheckin:
    properties:
      activity_id:
        type: integer
      checked_in_at:
        type: string
      created_at:
        type: string
      id:
        type: integer
      updated_at:
        type: string
      user:
        $ref: '#/definitions/entity.User'
      user_id:
        type: integer
    type: object
  entity.Address:
    properties:
      active:
        type: boolean
      address:
        type: string
      city:
        type: string
      country:
        type: string
      created_at:
        type: string
      customer_id:
        type: integer
      id:
        type: integer
      main:
        type: boolean
      name:
        description: SupplierID *uint          `json:"supplier_id"` // HAPUS field
          ini
        type: string
      postal_code:
        type: string
      state:
        type: string
      street:
        type: string
      updated_at:
        type: string
    type: object
  entity.Contact:
    properties:
      active:
        type: boolean
      birthdate:
        type: string
      created_at:
        type: string
      customer_id:
        type: integer
      department:
        type: string
      email:
        type: string
      id:
        type: integer
      job_position:
        type: string
      main:
        type: boolean
      mobile:
        type: string
      name:
        type: string
      phone:
        type: string
      position:
        type: string
      updated_at:
        type: string
    type: object
  entity.Customer:
    properties:
      account_manager_id:
        type: string
      activities:
        items:
          $ref: '#/definitions/entity.Activity'
        type: array
      addresses:
        description: Relations
        items:
          $ref: '#/definitions/entity.Address'
        type: array
      average_cost:
        type: number
      brand_name:
        type: string
      category:
        type: string
      code:
        type: string
      contacts:
        items:
          $ref: '#/definitions/entity.Contact'
        type: array
      created_at:
        type: string
      description:
        type: string
      email:
        type: string
      groups:
        items:
          $ref: '#/definitions/entity.Group'
        type: array
      id:
        type: integer
      logo:
        type: string
      name:
        type: string
      others:
        items:
          $ref: '#/definitions/entity.Other'
        type: array
      phone:
        type: string
      rating:
        type: number
      sosmeds:
        items:
          $ref: '#/definitions/entity.Sosmed'
        type: array
      status:
        description: Status internal
        type: string
      structures:
        items:
          $ref: '#/definitions/entity.Structure'
        type: array
      updated_at:
        type: string
      website:
        type: string
    type: object
  entity.Group:
    properties:
      active:
        type: boolean
      created_at:
        type: string
      customers:
        description: Relations
        items:
          $ref: '#/definitions/entity.Customer'
        type: array
      id:
        type: integer
      name_group:
        type: string
      updated_at:
        type: string
      value:
        type: string
    type: object
  entity.Other:
    properties:
      active:
        type: boolean
      created_at:
        type: string
      customer_id:
        type: integer
      id:
        type: integer
      key:
        type: string
      updated_at:
        type: string
      value:
        type: string
    type: object
  entity.Role:
    properties:
      created_at:
        type: string
      id:
        type: integer
      role_name:
        type: string
      updated_at:
        type: string
      users:
        description: Relations
        items:
          $ref: '#/definitions/entity.User'
        type: array
    type: object
  entity.Sosmed:
    properties:
      active:
        type: boolean
      created_at:
        type: string
      customer_id:
        type: integer
      followers:
        type: integer
      handle:
        type: string
      id:
        type: integer
      name:
        type: string
      platform:
        type: string
      updated_at:
        type: string
      url:
        type: string
      username:
        type: string
    type: object
  entity.Structure:
    properties:
      active:
        type: boolean
      address:
        type: string
      children:
        items:
          $ref: '#/definitions/entity.Structure'
        type: array
      created_at:
        type: string
      customer_id:
        type: integer
      id:
        type: integer
      level:
        type: integer
      name:
        type: string
      parent:
        $ref: '#/definitions/entity.Structure'
      parent_id:
        type: integer
      position:
        type: integer
      updated_at:
        type: string
    type: object
  entity.User:
    properties:
      activities:
        items:
          $ref: '#/definitions/entity.Activity'
        type: array
      activity_checkins:
        items:
          $ref: '#/definitions/entity.ActivityCheckin'
        type: array
      attending_activities:
        items:
          $ref: '#/definitions/entity.Activity'
        type: array
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      role:
        allOf:
        - $ref: '#/definitions/entity.Role'
        description: |-
          Update struct User, tambahkan di bagian Relations:
          Relations
      role_id:
        description: Default to regular user role
        type: integer
      updated_at:
        type: string
      username:
        type: string
    type: object
...
host: localhost:9000
info:
  contact: {}
  description: API untuk manajemen customer menggunakan Golang, Gin, dan PostgreSQL
  title: Customer Management API
  version: "1.0"
paths:
...
  /api/customers/{id}/others:
    get:
      consumes:
      - application/json
      description: Get all other attributes for specific customer
      parameters:
      - description: Customer ID
        in: path
        name: id
        required: true
        type: integer
      - description: Filter by active status
        in: query
        name: active
        type: boolean
      - description: Filter by attribute name
        in: query
        name: attribute_name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.Other'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get customer others
      tags:
      - Others
...
  /api/customers/{id}/with-others:
    get:
      consumes:
      - application/json
      description: Get customer data with all other attributes
      parameters:
      - description: Customer ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Customer'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get customer with others
      tags:
      - Customers
...
  /api/others:
    post:
      consumes:
      - application/json
      description: Create a new other field
      parameters:
      - description: Other data
        in: body
        name: other
        required: true
        schema:
          $ref: '#/definitions/dto.CreateOtherRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entity.Other'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create other
      tags:
      - Others
  /api/others/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an other attribute by ID
      parameters:
      - description: Other ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete other
      tags:
      - Others
    get:
      consumes:
      - application/json
      description: Get a specific other attribute by ID
      parameters:
      - description: Other ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Other'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get other by ID
      tags:
      - Others
  /api/others/by-attribute:
    get:
      consumes:
      - application/json
      description: Get all other attributes filtered by attribute name across all
        customers
      parameters:
      - description: Attribute name to filter
        in: query
        name: attribute_name
        required: true
        type: string
      - description: Filter by active status
        in: query
        name: active
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.Other'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get others by attribute name
      tags:
      - Others
  /api/roles:
    get:
      consumes:
      - application/json
      description: Get list of all roles
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.Role'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get all roles
      tags:
      - Roles
  /api/roles/{id}:
    get:
      consumes:
      - application/json
      description: Get a specific role by ID
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Role'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get role by ID
      tags:
      - Roles
...
securityDefinitions:
  BearerAuth:
    description: Type \"Bearer\" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"